ROM-117 ディレクトリ修正
