+++ src/components/parts/SecondaryButton/styles.ts
@@ -1,19 +1,18 @@
 import { css, CSSObject } from '@emotion/react';
 
+import {
+  DEFAULT_BUTTON_HEIGHT,
+  DEFAULT_BUTTON_WIDTH,
+} from '@/app/constants/sizes';
 import { borderRadiuses } from '@/app/theme/borderRadius';
-import { borderWidths } from '@/app/theme/borderWidths';
 import { colors } from '@/app/theme/colors';
-import { durations } from '@/app/theme/durations';
 import { scales } from '@/app/theme/scales';
 import { spaces } from '@/app/theme/spaces';
-
-import { DEFAULT_BUTTON_HEIGHT, DEFAULT_BUTTON_WIDTH } from '../constants';
 
 type ButtonSize = 'm' | 'l' | 'auto';
 
 export type StyleProps = {
   size?: ButtonSize;
-  isOutlined?: boolean;
 };
 
 const BUTTON_WIDTH: Record<ButtonSize, CSSObject['width']> = {
@@ -31,28 +30,22 @@ export const styles = {
   }),
 };
 
-export const getStyles = ({ size = 'auto', isOutlined }: StyleProps) => {
-  return css({
+export const getButtonStyle = ({ size = 'auto' }: StyleProps) =>
+  css({
     all: 'unset',
     overflow: 'hidden',
     whiteSpace: 'nowrap',
     height: DEFAULT_BUTTON_HEIGHT.MEDIUM,
     width: BUTTON_WIDTH[size],
-
+    borderRadius: borderRadiuses.m,
     paddingLeft: spaces.s,
     paddingRight: spaces.s,
 
-    borderRadius: borderRadiuses.m,
-    borderWidth: borderWidths.m,
-    borderColor: colors.tapa,
-    borderStyle: isOutlined ? 'solid' : 'none',
-
     ':enabled': {
+      backgroundColor: colors.mirage,
       cursor: 'pointer',
       ':hover': {
-        transitionDuration: durations.s,
-        borderColor: colors.black,
-        backgroundColor: colors.blackHaze,
+        backgroundColor: colors.black,
       },
       ':active': {
         transform: scales.smaller,
@@ -60,7 +53,7 @@ export const getStyles = ({ size = 'auto', isOutlined }: StyleProps) => {
     },
 
     ':disabled': {
+      backgroundColor: colors.blackHaze,
       cursor: 'not-allowed',
     },
   });
-};
